Changes made
Directly referencing issue #297 (Mediawiki output), I've added in the necessary functionality of integrating mediawiki tables as a potential output format. First, made the changes to get_formatted_string() function

       if out_format == "mediawiki": return self.get_mediawiki_string(**kwargs)

Next, I've added a new get_mediawiki_string() function that returns a string representation of the table in MediaWiki table markup.

I've also added in a from_mediawiki() function that takes in a multiline string of mediawiki markdown and converts it to a prettytable.

Lastly, I've also updated the documentation to reflect the changes.